+++ src/PockyBot.NET/Constants/Commands.cs
@@ -14,5 +14,6 @@ namespace PockyBot.NET.Constants
         public const string StringConfig = "stringconfig";
         public const string RemoveUser = "removeuser";
         public const string LocationWeight = "locationweight";
+        public const string Keywords = "keywords";
     }
 }